cleaned code and made few new files to keep everything in order. few new defines in ft_42sh.h to make it easier to read what is happening when calling either get_value or substring functions
