+++ sources/parameter_formatting/parameter_format.c
@@ -6,13 +6,11 @@
 /*   By: mviinika <mviinika@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 11:40:05 by mviinika          #+#    #+#             */
-/*   Updated: 2023/02/10 10:35:53 by mviinika         ###   ########.fr       */
+/*   Updated: 2023/02/10 11:53:10 by mviinika         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_42sh.h"
-
-extern t_shell *g_sh;
 
 static int check_syntax(char *cmd)
 {
@@ -26,240 +24,9 @@ static int check_syntax(char *cmd)
 	return (0);
 }
 
-static int is_legit(char *flag)
-{
-	if (*flag == '\0')
-		return (0);
-	else if (flag[0] == ':')
-		return (flag[1] == '-' || flag[1] == '+' || flag[1] == '=' || flag[1] == '?');
-	return ((flag[0] == '%' && flag[1]) || (flag[0] == '#' && flag[1]));
-}
 
-static char *get_operator(char *cmd, int *ret)
-{
-	int i;
 
-	i = 2;
-	(void)ret;
-	// ft_printf("getting operator %s\n",cmd);
-	if (!cmd[1])
-		return (NULL);
-	while (ft_isalnum(cmd[i]))
-		i++;
-	// ft_printf("getting operator %s\n",cmd + i);
-	if (!is_legit(&cmd[i]) && cmd[i] != '}')
-		*ret = -1;
-	return (cmd + i);
-}
-
-static char *get_flag(char *cmd, int *ret)
-{
-	int i;
-
-	i = 2;
-	(void)ret;
-	// ft_printf("getting flag %s\n",cmd);
-	if (!cmd[1])
-		return (NULL);
-	while (ft_isalnum(cmd[i]))
-		i++;
-	// if (!check_flag(cmd[i]))
-	// 	*ret = -1;
-	return (cmd + i);
-}
-
-static void init_pa(t_param *pa)
-{
-	pa->subs = NULL;
-	pa->strip = NULL;
-	pa->var = NULL;
-	pa->list = (char **)ft_memalloc(sizeof(char *) * 100);
-	pa->expanded = ft_strnew(1);
-	pa->op = 0;
-}
-
-static int splits(char *cmd, t_param *pa, int *ret)
-{
-	pa->oper = get_operator(cmd, ret);
-	if (is_legit(pa->oper))
-	{
-		pa->strip = ft_strdup(cmd);
-		remove_braces(pa->strip);
-		pa->subs = ft_strdup(get_flag(pa->strip + 1, ret));
-		//"%s\n", pa->subs);
-		pa->op = pa->subs[0];
-		pa->subs =
-			(char *)ft_memmove(pa->subs, pa->subs + 1, ft_strlen(pa->subs));
-		pa->var = ft_strndup(pa->strip, ft_strlen(pa->strip) - (ft_strlen(pa->subs)));
-		// ft_printf("expanded [%s]: \n", pa->subs);
-	}
-	else if (*ret == -1)
-	{
-		ft_printf("42sh: %s: bad substitution\n", cmd);
-		return (1);
-	}
-	else
-	{
-		ft_strdel(&pa->expanded);
-		pa->strip = ft_strdup(cmd);
-		remove_braces(pa->strip);
-		pa->expanded = ft_expansion_dollar(g_sh, pa->strip);
-		ft_strdel(&pa->strip);
-		*ret = 2;
-		return (1);
-	}
-	return (0);
-}
-
-static void retoken_into_list(t_param *pa)
-{
-	int i;
-	int j;
-
-	i = 0;
-	j = 0;
-	while (pa->subs[j])
-	{
-		pa->list[i++] = retokenize(pa->subs, &j);
-	}
-}
-
-char *variable_length(char *str)
-{
-	char *var;
-	char *temp;
-	int i;
-	int k;
-
-	i = 0;
-	k = 0;
-	var = ft_strnew(ft_strlen(str));
-	while (str[i])
-	{
-		var[k++] = str[i++];
-		if (str[i] == '#')
-			i++;
-	}
-	temp = ft_expansion_dollar(g_sh, var);
-	i = ft_strlen(temp);
-	return (ft_itoa(i));
-}
-
-static int expander(t_param *pa, int ret)
-{
-	int i;
-	char *temp;
-	char *subs;
-
-	i = -1;
-	while (pa->list[++i])
-	{
-		pa->flag = get_flag(pa->list[i], &ret);
-		// if (!pa->flag)
-		// 	pa->expanded = ft_strnew(1);
-		// ft_printf("pa list %s  [%c]\n",  pa->list[i], pa->flag[0]);
-		if (ft_strnequ(pa->list[i], "${", 2) && pa->flag[0] == ':')
-		{
-			subs = substitute_or_create(g_sh, pa->list[i], &ret);
-			temp = ft_strjoin(pa->expanded, subs);
-			ft_strdel(&pa->expanded);
-			if (!temp || !*temp)
-			{
-				pa->expanded = ft_strdup(" ");
-				ft_strdel(&subs);
-				ft_strdel(&temp);
-				return (1);
-			}
-				
-			pa->expanded = ft_strdup(temp);
-			ft_strdel(&subs);
-			ft_strdel(&temp);
-		}
-		else if ((ft_strnequ(pa->list[i], "${", 2) && pa->flag[0] == '#') || (ft_strnequ(pa->list[i], "${", 2) && pa->flag[0] == '%'))
-		{
-			ft_printf(" eju [%s] \n", pa->list[i]);
-			subs = search_from_var(g_sh, pa->list[i], &ret);
-			temp = ft_strjoin(pa->expanded, subs);
-			
-			ft_strdel(&pa->expanded);
-			pa->expanded = ft_strdup(temp);
-			// ft_printf(" expanded [%s] \n", pa->expanded);
-			ft_strdel(&subs);
-			ft_strdel(&temp);
-		}
-		else
-			pa->expanded = ft_strupdate(pa->expanded, pa->list[i]);
-		if (ret == -1 || !*pa->expanded)
-			return (1);
-	}
-	return (0);
-}
-
-static int joiner(t_shell *sh, t_param *pa, char *cmd, int ret)
-{
-	char *temp;
-	char *subs;
-
-	ft_strdel(&pa->subs);
-	pa->subs = ft_strjoin(pa->var, pa->expanded);
-	pa->oper = get_operator(pa->subs, &ret);
-	// ft_printf("pa list sdfds %s\n", pa->subs);
-	// ft_printf("pa list  %c\n", pa->flag[0]);
-	ft_strdel(&pa->expanded);
-	pa->expanded = ft_strnew(1);
-	if (pa->oper[0] == ':')
-	{
-		subs = substitute_or_create(g_sh, pa->subs, &ret);
-		temp = ft_strjoin(pa->expanded, subs);
-		if (temp == NULL)
-			return (1);
-		ft_strdel(&pa->expanded);
-		pa->expanded = ft_strdup(temp);
-		ft_strdel(&subs);
-		ft_strdel(&temp);
-		// pa->expanded = ft_strupdate(pa->expanded,
-		// substitute_or_create(g_sh, pa->subs, &ret));
-	}
-	else if (pa->oper[0] == '#' || pa->oper[0] == '%')
-	{
-		subs = search_from_var(g_sh, pa->subs, &ret);
-		if (!subs)
-			return (1);
-		// ft_printf(" subs [%s] \n", subs);
-		temp = ft_strjoin(pa->expanded, subs);
-		ft_strdel(&pa->expanded);
-		pa->expanded = ft_strdup(temp);
-		// ft_printf(" expanded [%s] \n", pa->expanded);
-		ft_strdel(&subs);
-		ft_strdel(&temp);
-		// pa->expanded = ft_strupdate(pa->expanded,
-		// search_from_var(g_sh, pa->subs, &ret));
-	}
-	else if (pa->subs[0] == '$' || pa->subs[1] == '#')
-	{
-		pa->expanded = variable_length(pa->subs);
-		// ft_printf("%s\n", pa->expanded);
-		return (0);
-	}
-	else
-		pa->expanded = ft_expansion_dollar(sh, cmd);
-	if (!pa->expanded || !*pa->expanded)
-	{
-		pa->expanded = ft_strnew(1);
-		// ft_printf("42sh: %s: bad substitution\n", cmd);
-		//  return (-1);
-	}
-	if (ret == -1)
-	{
-		ft_strdel(&pa->expanded);
-		ft_printf("42sh: %s: bad substitution\n", cmd);
-		return (-1);
-	}
-	// ft_printf("pa list alaalla %s\n", pa->expanded);
-	return (0);
-}
-
-char *separate_form(char *cmd)
+static char *separate_form(char *cmd)
 {
 	char *fresh;
 	int i;
@@ -285,19 +52,7 @@ char *separate_form(char *cmd)
 	return (fresh);
 }
 
-void	free_attrs(t_param *pa, char **new_cmd)
-{
-	int	k;
-
-	k = 0;
-	while (k < 100)
-		ft_strdel(&pa->list[k++]);
-	free(pa->list);
-	ft_strdel(&pa->expanded);
-	ft_strdel(&(*new_cmd));
-}
-
-char	*init_form_cycle(t_param *pa, char *cmd, t_pa_ints *ints)
+static char	*init_form_cycle(t_param *pa, char *cmd, t_pa_ints *ints)
 {
 	char *new_cmd;
 	init_pa(pa);
@@ -309,37 +64,6 @@ char	*init_form_cycle(t_param *pa, char *cmd, t_pa_ints *ints)
 		ints->j += ft_strlen(new_cmd);
 	}
 	return (new_cmd);
-}
-
-int	perform_param_expans(char *cmd, t_param *pa, int *ret)
-{
-	int	err;
-
-	err = 0;
-	if (splits(cmd, pa, ret) && *ret == 0)
-	{
-		;
-	}
-	if (*ret == 0)
-		retoken_into_list(pa);
-	if (expander(pa, *ret))
-		err = 1;
-	if (*ret == 0 && joiner(g_sh, pa, cmd, *ret) && *ret == 0)
-		err = -1;
-	if (*ret == 1)
-	{
-		err = -1;
-	}
-	return (err);
-}
-
-void init_pa_ints(t_pa_ints *ints, char **new_cmd)
-{
-	ints->err = 0;
-	ints->i = -1;
-	ints->j = 0;
-	ints->ret = 0;
-	(*new_cmd) = NULL;
 }
 
 int param_format(char **cmd)
